All pages linked and using complete w3 style sheet
These are all the pages to the website using the w3 CSS style sheet. I included these copies because I think w3css is useful and I do want to use other code that is included in the style sheet. However, I am not trying to ONLY use this stylesheet because I need to learn how to do all the various things on my own. Using it as more of a guide to teach myself how to do what is included on the stylesheet.